エフェクトマネージャのヌル参照バグを修正

+++ Project/App/Feature/Pilot/Mech/Weapon/MachineGun/PilotMechWeaponMachineGun.cpp
@@ -89,7 +89,7 @@ void PilotMechWeaponMachineGun::Attack([[maybe_unused]] Damage::Power power) {
 		PilotMech* pMech = dynamic_cast<PilotMech*>(mech_);
 		const float attackMul = pMech->GetStatus()->GetAttackMul();
 		const float damageF = damage * attackMul;
-		param.initParam_.gParam.damage = damageF;
+		param.initParam_.gParam.damage = static_cast<int32_t>(damageF);
 		param.initParam_.gParam.power = Damage::Power::Small;
 		param.initParam_.gParam.tag = FriendlyTag::PlayerSide;
